fix: 프로젝트 룰에 따른 파일 재배치
