--- src/DummyData/cardData.tsx
@@ -1,199 +1,0 @@
-interface IImageContain {
-  webSrcSet: string;
-  imgSrc: string;
-  alt: string;
-}
-
-interface Card {
-  ImageContain: IImageContain[];
-}
-
-export const cardData: Card[] = [
-  {
-    ImageContain: [
-      {
-        webSrcSet:
-          "https://image.wavve.com/v1/thumbnails/240_360_20_80/meta/image/202311/1699601487098289299.webp 1x,https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202311/1699601487098289299.webp 1.7x,https://image.wavve.com/v1/thumbnails/720_1080_20_80/meta/image/202311/1699601487098289299.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202311/1699601487098289299.jpg",
-        alt: "나 혼자 산다",
-      },
-    ],
-  },
-  {
-    ImageContain: [
-      {
-        webSrcSet:
-          "https://image.wavve.com/v1/thumbnails/240_360_20_80/meta/image/202311/1698908219745378929.webp 1x,https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202311/1698908219745378929.webp 1.7x,https://image.wavve.com/v1/thumbnails/720_1080_20_80/meta/image/202311/1698908219745378929.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202311/1698908219745378929.jpg",
-        alt: "런닝맨",
-      },
-    ],
-  },
-  {
-    ImageContain: [
-      {
-        webSrcSet:
-          "https://image.wavve.com/v1/thumbnails/240_360_20_80/meta/image/202402/1708404301458518711.webp 1x,https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202402/1708404301458518711.webp 1.7x,https://image.wavve.com/v1/thumbnails/720_1080_20_80/meta/image/202402/1708404301458518711.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202402/1708404301458518711.jpg",
-        alt: "골 때리는 그녀들",
-      },
-    ],
-  },
-  {
-    ImageContain: [
-      {
-        webSrcSet:
-          "https://image.wavve.com/v1/thumbnails/240_360_20_80/meta/image/202407/1719877649144513611.webp 1x,https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202407/1719877649144513611.webp 1.7x,https://image.wavve.com/v1/thumbnails/720_1080_20_80/meta/image/202407/1719877649144513611.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202407/1719877649144513611.jpg",
-        alt: "모든패밀리",
-      },
-    ],
-  },
-  {
-    ImageContain: [
-      {
-        webSrcSet:
-          "https://image.wavve.com/v1/thumbnails/240_360_20_80/meta/image/202311/1699601487098289299.webp 1x,https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202311/1699601487098289299.webp 1.7x,https://image.wavve.com/v1/thumbnails/720_1080_20_80/meta/image/202311/1699601487098289299.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202311/1699601487098289299.jpg",
-        alt: "나 혼자 산다",
-      },
-    ],
-  },
-  {
-    ImageContain: [
-      {
-        webSrcSet:
-          "https://image.wavve.com/v1/thumbnails/240_360_20_80/meta/image/202311/1698908219745378929.webp 1x,https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202311/1698908219745378929.webp 1.7x,https://image.wavve.com/v1/thumbnails/720_1080_20_80/meta/image/202311/1698908219745378929.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202311/1698908219745378929.jpg",
-        alt: "런닝맨",
-      },
-    ],
-  },
-  {
-    ImageContain: [
-      {
-        webSrcSet:
-          "https://image.wavve.com/v1/thumbnails/240_360_20_80/meta/image/202402/1708404301458518711.webp 1x,https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202402/1708404301458518711.webp 1.7x,https://image.wavve.com/v1/thumbnails/720_1080_20_80/meta/image/202402/1708404301458518711.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202402/1708404301458518711.jpg",
-        alt: "골 때리는 그녀들",
-      },
-    ],
-  },
-  {
-    ImageContain: [
-      {
-        webSrcSet:
-          "https://image.wavve.com/v1/thumbnails/240_360_20_80/meta/image/202407/1719877649144513611.webp 1x,https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202407/1719877649144513611.webp 1.7x,https://image.wavve.com/v1/thumbnails/720_1080_20_80/meta/image/202407/1719877649144513611.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202407/1719877649144513611.jpg",
-        alt: "모든패밀리",
-      },
-    ],
-  },
-  {
-    ImageContain: [
-      {
-        webSrcSet:
-          "https://image.wavve.com/v1/thumbnails/240_360_20_80/meta/image/202311/1699601487098289299.webp 1x,https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202311/1699601487098289299.webp 1.7x,https://image.wavve.com/v1/thumbnails/720_1080_20_80/meta/image/202311/1699601487098289299.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202311/1699601487098289299.jpg",
-        alt: "나 혼자 산다",
-      },
-    ],
-  },
-  {
-    ImageContain: [
-      {
-        webSrcSet:
-          "https://image.wavve.com/v1/thumbnails/240_360_20_80/meta/image/202311/1698908219745378929.webp 1x,https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202311/1698908219745378929.webp 1.7x,https://image.wavve.com/v1/thumbnails/720_1080_20_80/meta/image/202311/1698908219745378929.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202311/1698908219745378929.jpg",
-        alt: "런닝맨",
-      },
-    ],
-  },
-  {
-    ImageContain: [
-      {
-        webSrcSet:
-          "https://image.wavve.com/v1/thumbnails/240_360_20_80/meta/image/202402/1708404301458518711.webp 1x,https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202402/1708404301458518711.webp 1.7x,https://image.wavve.com/v1/thumbnails/720_1080_20_80/meta/image/202402/1708404301458518711.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202402/1708404301458518711.jpg",
-        alt: "골 때리는 그녀들",
-      },
-    ],
-  },
-  {
-    ImageContain: [
-      {
-        webSrcSet:
-          "https://image.wavve.com/v1/thumbnails/240_360_20_80/meta/image/202407/1719877649144513611.webp 1x,https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202407/1719877649144513611.webp 1.7x,https://image.wavve.com/v1/thumbnails/720_1080_20_80/meta/image/202407/1719877649144513611.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202407/1719877649144513611.jpg",
-        alt: "모든패밀리",
-      },
-    ],
-  },
-  {
-    ImageContain: [
-      {
-        webSrcSet:
-          "https://image.wavve.com/v1/thumbnails/240_360_20_80/meta/image/202311/1699601487098289299.webp 1x,https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202311/1699601487098289299.webp 1.7x,https://image.wavve.com/v1/thumbnails/720_1080_20_80/meta/image/202311/1699601487098289299.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202311/1699601487098289299.jpg",
-        alt: "나 혼자 산다",
-      },
-    ],
-  },
-  {
-    ImageContain: [
-      {
-        webSrcSet:
-          "https://image.wavve.com/v1/thumbnails/240_360_20_80/meta/image/202311/1698908219745378929.webp 1x,https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202311/1698908219745378929.webp 1.7x,https://image.wavve.com/v1/thumbnails/720_1080_20_80/meta/image/202311/1698908219745378929.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202311/1698908219745378929.jpg",
-        alt: "런닝맨",
-      },
-    ],
-  },
-  {
-    ImageContain: [
-      {
-        webSrcSet:
-          "https://image.wavve.com/v1/thumbnails/240_360_20_80/meta/image/202402/1708404301458518711.webp 1x,https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202402/1708404301458518711.webp 1.7x,https://image.wavve.com/v1/thumbnails/720_1080_20_80/meta/image/202402/1708404301458518711.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202402/1708404301458518711.jpg",
-        alt: "골 때리는 그녀들",
-      },
-    ],
-  },
-  {
-    ImageContain: [
-      {
-        webSrcSet:
-          "https://image.wavve.com/v1/thumbnails/240_360_20_80/meta/image/202407/1719877649144513611.webp 1x,https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202407/1719877649144513611.webp 1.7x,https://image.wavve.com/v1/thumbnails/720_1080_20_80/meta/image/202407/1719877649144513611.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/480_720_20_80/meta/image/202407/1719877649144513611.jpg",
-        alt: "모든패밀리",
-      },
-    ],
-  },
-];
-
-export const NAME_CONTAIN: string[] = [
-  "믿고 보는 웨이브 에디터 추천작",
-  "실시간 인기 콘텐츠",
-  "오직 웨이브에서",
-  "새로 올라온 시리즈",
-  "새로 올라온 영화",
-  "제 28회 부천국제판타스틱 영화제",
-  "부담 없이 보는 단막극",
-  "업고 튀고 싶은 최애 하나쯤 있잖아",
-];
