--- src/DummyData/slideData.tsx
@@ -1,94 +1,0 @@
-// slideCtx.ts
-interface Image {
-  webpSrcSet: string;
-  imgSrc: string;
-  alt: string;
-  className?: string;
-}
-
-interface Slide {
-  images: Image[];
-}
-
-export const slideData: Slide[] = [
-  {
-    images: [
-      {
-        webpSrcSet:
-          "https://image.wavve.com/v1/thumbnails/1240_508_20_80/banner/pooq/2023/20230901_banner_152933.webp 1x, https://image.wavve.com/v1/thumbnails/2480_1016_20_80/banner/pooq/2023/20230901_banner_152933.webp 1.7x, https://image.wavve.com/v1/thumbnails/3720_1524_20_80/banner/pooq/2023/20230901_banner_152933.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/2480_1016_20_80/banner/pooq/2023/20230901_banner_152933.jpg",
-        alt: "wavvie영화_[감독판] 윈드 리버",
-        className: "big-banner",
-      },
-      {
-        webpSrcSet:
-          "https://image.wavve.com/v1/thumbnails/1240_508_20_80/banner/pooq/2023/imgbuild_20230901_153033694.webp 1x, https://image.wavve.com/v1/thumbnails/2480_1016_20_80/banner/pooq/2023/imgbuild_20230901_153033694.webp 1.7x, https://image.wavve.com/v1/thumbnails/3720_1524_20_80/banner/pooq/2023/imgbuild_20230901_153033694.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/2480_1016_20_80/banner/pooq/2023/imgbuild_20230901_153033694.png",
-        alt: "wavvie영화_[감독판] 윈드 리버",
-        className: "banner-title",
-      },
-    ],
-  },
-  {
-    images: [
-      {
-        webpSrcSet:
-          "https://image.wavve.com/v1/thumbnails/1240_508_20_80/banner/pooq/2024/20240701_banner_160545.webp 1x,https://image.wavve.com/v1/thumbnails/2480_1016_20_80/banner/pooq/2024/20240701_banner_160545.webp 1.7x,https://image.wavve.com/v1/thumbnails/3720_1524_20_80/banner/pooq/2024/20240701_banner_160545.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/2480_1016_20_80/banner/pooq/2024/20240701_banner_160545.jpg",
-        alt: "셋",
-        className: "big-banner",
-      },
-      {
-        webpSrcSet:
-          "https://image.wavve.com/v1/thumbnails/1240_508_20_80/banner/pooq/2024/imgbuild_20240701_161531414.webp 1x,https://image.wavve.com/v1/thumbnails/2480_1016_20_80/banner/pooq/2024/imgbuild_20240701_161531414.webp 1.7x,https://image.wavve.com/v1/thumbnails/3720_1524_20_80/banner/pooq/2024/imgbuild_20240701_161531414.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/2480_1016_20_80/banner/pooq/2024/imgbuild_20240701_161531414.png",
-        alt: "wavvie영화_[감독판] 윈드 리버",
-        className: "banner-title",
-      },
-    ],
-  },
-  {
-    images: [
-      {
-        webpSrcSet:
-          "https://image.wavve.com/v1/thumbnails/3720_1524_20_80/banner/pooq/2024/20240507_banner_112843.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/2480_1016_20_80/banner/pooq/2024/20240507_banner_112843.jpg",
-        alt: "애니_아따맘마 리마스터",
-        className: "big-banner",
-      },
-      {
-        webpSrcSet:
-          "https://image.wavve.com/v1/thumbnails/1240_508_20_80/banner/pooq/2024/imgbuild_20240510_180656890.webp 1x,https://image.wavve.com/v1/thumbnails/2480_1016_20_80/banner/pooq/2024/imgbuild_20240510_180656890.webp 1.7x,https://image.wavve.com/v1/thumbnails/3720_1524_20_80/banner/pooq/2024/imgbuild_20240510_180656890.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/2480_1016_20_80/banner/pooq/2024/imgbuild_20240510_180656890.png",
-        alt: "wavvie영화_[감독판] 윈드 리버",
-        className: "banner-title",
-      },
-    ],
-  },
-  {
-    images: [
-      {
-        webpSrcSet:
-          "https://image.wavve.com/v1/thumbnails/1240_508_20_80/banner/pooq/2024/20240704_banner_100411.webp 1x,https://image.wavve.com/v1/thumbnails/2480_1016_20_80/banner/pooq/2024/20240704_banner_100411.webp 1.7x,https://image.wavve.com/v1/thumbnails/3720_1524_20_80/banner/pooq/2024/20240704_banner_100411.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/2480_1016_20_80/banner/pooq/2024/20240704_banner_100411.jpg",
-        alt: "asd",
-        className: "big-banner",
-      },
-      {
-        webpSrcSet:
-          "https://image.wavve.com/v1/thumbnails/1240_508_20_80/banner/pooq/2024/imgbuild_20240704_101804428.webp 1x,https://image.wavve.com/v1/thumbnails/2480_1016_20_80/banner/pooq/2024/imgbuild_20240704_101804428.webp 1.7x,https://image.wavve.com/v1/thumbnails/3720_1524_20_80/banner/pooq/2024/imgbuild_20240704_101804428.webp 2.5x",
-        imgSrc:
-          "https://image.wavve.com/v1/thumbnails/2480_1016_20_80/banner/pooq/2024/imgbuild_20240704_101804428.png",
-        alt: "wavvie영화_[감독판] 윈드 리버",
-        className: "banner-title",
-      },
-    ],
-  },
-];
