+++ src/components/common/footer/index.tsx
@@ -1,4 +1,5 @@
 const Footer = () => {
-  return <footer>footer</footer>;
+    return <footer>footer</footer>;
 };
+
 export default Footer;
