+++ src/components/layout/style.ts
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const MainContainer = styled.main``;
+
+export const HeaderContainer = styled.header``;
+
+export const FooterContainer = styled.footer``;
