+++ src/constants/nav-menus.constant.ts
@@ -1,0 +1,21 @@
+export const GNB_MENUS_LIST: string[] = [
+    "드라마",
+    "영화",
+    "애니",
+    "해외시리즈",
+    "시사교양",
+    "키즈",
+    "영화플러스",
+    "LIVE",
+];
+
+export const CAROUSEL_TITLE_LIST: string[] = [
+    "믿고 보는 웨이브 에디터 추천작",
+    "실시간 인기 콘텐츠",
+    "오직 웨이브에서",
+    "새로 올라온 시리즈",
+    "새로 올라온 영화",
+    "제 28회 부천국제판타스틱 영화제",
+    "부담 없이 보는 단막극",
+    "업고 튀고 싶은 최애 하나쯤 있잖아",
+];
