+++ src/index.tsx
@@ -1,15 +1,15 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
-import { ThemeModeProvider } from "./components/Theme/themeCtx";
+import { ThemeModeProvider } from "./contexts/themeCtx";
 
 const root = ReactDOM.createRoot(
-  document.getElementById("root") as HTMLElement
+    document.getElementById("root") as HTMLElement
 );
 root.render(
-  <React.StrictMode>
-    <ThemeModeProvider>
-      <App />
-    </ThemeModeProvider>
-  </React.StrictMode>
+    <React.StrictMode>
+        <ThemeModeProvider>
+            <App />
+        </ThemeModeProvider>
+    </React.StrictMode>
 );
